requirements.txt
voluptuous>=0.13.0,<1
PyYAML>=6.0,<7
colorama==0.4.4
